Crafting Contraptions (#63)
Closes #62

Added crafting UI to handle the crafting of contraptions as well as the the api function triggerContraption() for contraptions to call when used.
Changes were also made to unusual materials.

Also added helper promptItemUse() which prompts users to decide if they wish to use an item or only post the item's card into chat.